Moved to separate examples sr directory
